+++ src/instances/if-then-else.js
@@ -256,6 +256,10 @@ class InstanceIfThenElse extends Instance {
     return fittestIndex
   }
 
+  hasNestedValidationErrors () {
+    return this.activeInstance ? this.activeInstance.hasNestedValidationErrors() : false
+  }
+
   destroy () {
     this.instances.forEach((instance) => {
       instance.destroy()